feat(celestial): complete modular decomposition of celestial object system
Transform monolithic celestial.ts (1,761 lines) into focused modular architecture:

**New Modular Structure:**
- CelestialTypes.ts (97 lines) - Shared interfaces, base CelestialObject class
- Star.ts (494 lines) - Complete star system with visual effects, corona, sunspots
- Planet.ts (771 lines) - Complex planet system with atmospheres, rings, weather
- Moon.ts (167 lines) - Moon system with orbital mechanics
- celestial.ts (22 lines) - Clean barrel export maintaining backward compatibility

**Key Architectural Improvements:**
✅ Resolved circular dependencies between Star ↔ Planet ↔ Moon classes
✅ Added missing updatePosition method to Planet class for orbital mechanics
✅ Proper TypeScript interfaces with type safety throughout
✅ Single responsibility principle - each class focused on one object type
✅ Maintained 100% backward compatibility - existing imports unchanged

**Complex Systems Preserved:**
- Star visual effects: corona, sunspots, radiation, shimmer, swirling patterns
- Planet atmospheres: multi-layer atmospheric rendering with auroras and weather
- Planet ring systems: multi-band rings with shimmer effects for gas giants
- Planet surface features: craters, stripes, swirls, dunes, crystalline, lava flows
- Moon orbital mechanics: deterministic generation and parent planet relationships
- 7 Planet types (Rocky, Ocean, Gas Giant, Desert, Frozen, Volcanic, Exotic)
- 7 Star types (G-Type, K-Type, M-Type, Red Giant, Blue Giant, White Dwarf, Neutron)

**Quality Assurance:**
✅ All 774 tests passing throughout extraction process
✅ TypeScript compilation successful with zero errors
✅ Zero functionality regressions - all game features preserved
✅ Proper module dependency resolution and imports

This completes the celestial system portion of Phase 2 monolithic file decomposition.
The codebase now has clean, maintainable, and extensible celestial object architecture.
